+++ packages/widgets/lib/src/main.reflectable.dart
@@ -894,9 +894,10 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix26.IconIntrinsicStates() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix26.IconIntrinsicStates() : null,
               r'create': (bool b) =>
-                  () => b ? prefix26.IconIntrinsicStates.create() : null
+                  () => b ? const prefix26.IconIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -945,9 +946,9 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix26.ListViewIntrinsicStates() : null,
-              r'create': (bool b) =>
-                  () => b ? prefix26.ListViewIntrinsicStates.create() : null
+                  () => b ? const prefix26.ListViewIntrinsicStates() : null,
+              r'create': (bool b) => () =>
+                  b ? const prefix26.ListViewIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -1023,9 +1024,9 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix26.ScaffoldIntrinsicStates() : null,
-              r'create': (bool b) =>
-                  () => b ? prefix26.ScaffoldIntrinsicStates.create() : null
+                  () => b ? const prefix26.ScaffoldIntrinsicStates() : null,
+              r'create': (bool b) => () =>
+                  b ? const prefix26.ScaffoldIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -1048,9 +1049,10 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix26.StackIntrinsicStates() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix26.StackIntrinsicStates() : null,
               r'create': (bool b) =>
-                  () => b ? prefix26.StackIntrinsicStates.create() : null
+                  () => b ? const prefix26.StackIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -1179,7 +1181,10 @@ final _data = <r.Reflectable, r.ReflectorData>{
             -1,
             {},
             {},
-            {r'': (bool b) => () => b ? prefix27.WidgetAdapterParse() : null},
+            {
+              r'': (bool b) =>
+                  () => b ? const prefix27.WidgetAdapterParse() : null
+            },
             -1,
             -1,
             const <int>[-1],
@@ -1226,7 +1231,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.AlignWidgetAdapter() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix28.AlignWidgetAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.AlignWidgetAdapter.create() : null
             },
@@ -1258,7 +1264,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.ButtonAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.ButtonAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.ButtonAdapter.create() : null
             },
@@ -1290,7 +1296,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.ColumnAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.ColumnAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.ColumnAdapter.create() : null
             },
@@ -1322,7 +1328,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.BoxAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.BoxAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.BoxAdapter.create() : null
             },
@@ -1354,7 +1360,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.GridViewAdapter() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix28.GridViewAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.GridViewAdapter.create() : null
             },
@@ -1386,7 +1393,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.MaterialIconAdapter() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix28.MaterialIconAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.MaterialIconAdapter.create() : null
             },
@@ -1418,7 +1426,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.ImageAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.ImageAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.ImageAdapter.create() : null
             },
@@ -1450,7 +1458,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.ListViewAdapter() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix28.ListViewAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.ListViewAdapter.create() : null
             },
@@ -1482,7 +1491,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.LottieAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.LottieAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.LottieAdapter.create() : null
             },
@@ -1514,7 +1523,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.RowAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.RowAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.RowAdapter.create() : null
             },
@@ -1546,7 +1555,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.ScaffoldAdapter() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix28.ScaffoldAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.ScaffoldAdapter.create() : null
             },
@@ -1578,7 +1588,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.StackAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.StackAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.StackAdapter.create() : null
             },
@@ -1610,7 +1620,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.TextAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.TextAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.TextAdapter.create() : null
             },
@@ -1642,7 +1652,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.TextFieldAdapter() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix28.TextFieldAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.TextFieldAdapter.create() : null
             },
@@ -1674,7 +1685,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.VideoAdapter() : null,
+              r'': (bool b) => () => b ? const prefix28.VideoAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.VideoAdapter.create() : null
             },
@@ -1706,7 +1717,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix28.ComponentAdapter() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix28.ComponentAdapter() : null,
               r'create': (bool b) =>
                   () => b ? prefix28.ComponentAdapter.create() : null
             },
@@ -1780,8 +1792,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) =>
-                  () => b ? prefix30.AlignWidgetDefaultAttributes() : null,
+              r'': (bool b) => () =>
+                  b ? const prefix30.AlignWidgetDefaultAttributes() : null,
               r'create': (bool b) => () =>
                   b ? prefix30.AlignWidgetDefaultAttributes.create() : null
             },
@@ -1814,7 +1826,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.ButtonDefaultAttributes() : null,
+                  () => b ? const prefix30.ButtonDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.ButtonDefaultAttributes.create() : null
             },
@@ -1847,7 +1859,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.FlexDefaultAttributes() : null,
+                  () => b ? const prefix30.FlexDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.FlexDefaultAttributes.create() : null
             },
@@ -1881,7 +1893,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.ComponentDefaultAttributes() : null,
+                  () => b ? const prefix30.ComponentDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.ComponentDefaultAttributes.create() : null
             },
@@ -1913,7 +1925,8 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix30.BoxDefaultAttributes() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix30.BoxDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.BoxDefaultAttributes.create() : null
             },
@@ -1947,7 +1960,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.IconDefaultAttributes() : null,
+                  () => b ? const prefix30.IconDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.IconDefaultAttributes.create() : null
             },
@@ -1980,7 +1993,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.ListViewDefaultAttributes() : null,
+                  () => b ? const prefix30.ListViewDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.ListViewDefaultAttributes.create() : null
             },
@@ -2013,7 +2026,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.LottieDefaultAttributes() : null,
+                  () => b ? const prefix30.LottieDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.LottieDefaultAttributes.create() : null
             },
@@ -2046,7 +2059,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.ScaffoldDefaultAttributes() : null,
+                  () => b ? const prefix30.ScaffoldDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.ScaffoldDefaultAttributes.create() : null
             },
@@ -2079,7 +2092,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.StackDefaultAttributes() : null,
+                  () => b ? const prefix30.StackDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.StackDefaultAttributes.create() : null
             },
@@ -2112,7 +2125,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.TextDefaultAttributes() : null,
+                  () => b ? const prefix30.TextDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.TextDefaultAttributes.create() : null
             },
@@ -2145,7 +2158,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.TextFieldDefaultAttributes() : null,
+                  () => b ? const prefix30.TextFieldDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.TextFieldDefaultAttributes.create() : null
             },
@@ -2178,7 +2191,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix30.VideoDefaultAttributes() : null,
+                  () => b ? const prefix30.VideoDefaultAttributes() : null,
               r'create': (bool b) =>
                   () => b ? prefix30.VideoDefaultAttributes.create() : null
             },
@@ -2246,7 +2259,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix29.DefaultAttributesParse() : null
+                  () => b ? const prefix29.DefaultAttributesParse() : null
             },
             -1,
             -1,
@@ -3183,7 +3196,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             -1,
             {},
             {},
-            {r'': (bool b) => () => b ? prefix31.NodesParse() : null},
+            {r'': (bool b) => () => b ? const prefix31.NodesParse() : null},
             -1,
             -1,
             const <int>[-1],
@@ -3202,7 +3215,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
       ])
 };
 
-final _memberSymbolMap = null;
+const _memberSymbolMap = null;
 
 void initializeReflectable() {
   r.data = _data;